trimit_requirements.txt
bcrypt==3.1.5
certifi==2018.11.29
cffi==1.11.5
Django==1.11.17
django-countries==5.3.3
django-tagulous==0.14.0
geographiclib==1.49
geopy==1.18.1
olefile==0.46
Pillow==5.3.0
pycparser==2.19
pytz==2018.7
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
wincertstore==0.2
